Terminate any existing IdP session before SSO
When multiple projects exist, this breaks the ability of a user to log
out of AWS and use SPIN to sign in to a different project (sometimes).

Shibboleth IdP allows a session to have multiple principal names if they
have authenticated multiple times. The order is seemingly random so we
aren't even guaranteed to see the most recent principal name. This means
we can end up signing people into the wrong project.