Enhance the Project model and associated UI
Previously the project model had a text field called 'state' which was
not of use for its intended purpose of indicating if the project was
currently active or not with non active projects not being eligible
for session creation at AWS.

The field state is renamed to active in this commit and changed to
boolean as was originally intended.

Associated UI and API changes also accompany this commit.

Finally UI change to precisely describe 'Provider ARN' to SPIN
administrators is wrapped into this commit.

Closes #25.