Only set SSLVerifyClient at /api path
When SSLVerifyClient is configued at the top level 90% of awesome
browsers behave correctly and don't prompt for the user to supply a
cert to complete the request and then there is.... Safari.

Which does not do this and asks the user to supply a cert :panda_face:.

By specifying this directive within the /api Location only Apache will
renegotiate before completing api requests and not specify anything
about client SSL at the top level correcting the problem in Safari.