Fix an issue with merging an incomplete Subject
This fix relates to the unlikely scenario of a user with two incomplete
Subject records - one with a federated id (e.g. shared_token) populated,
and one which was created by invitation and therefore a candidate for
merge into the former.

When the invitation is accepted, the flow of events should be:

1. Resolve the incomplete Subject record which has a matching
   shared_token
2. Complete that Subject record using the authenticated details
3. Proceed with accepting the invitation
4. Merge the other incomplete Subject into the one completed in step 2

The existing logic failed to find the existing incomplete record, and
also was unable to complete a subject while in the "Accept Invitation"
workflow.