Handled special case if invitation mail exists
Subjects may exist who have a different email address to their invitation.
There is no way to find this email address in the system once the invitation
has been accepted. It might be confusing for the administrator to re-use this
email address in an invite, so an appropriate error message has been created.