Fixed special case when role revoked on invitation
It's possible for the administrator to revoke a role from a user after they've
been invited. The expired/show pages for the invitation were generating 500
server error — this was caused by the assumption that the Subject's first
ProjectRole always existed. Now there are checks around this.

The logic of a user's "first project" via invite has been strengthened to
ensure the project is active too.

As the system was standing, if the project was made inactive, the user would
still see the project referenced when viewing the invitation. Once they logged
in they were halted at the dashboard for not having a project. Now for this
case, the invitation expired/show views will not list a specific project.

The error message displayed on the dashboard has also been updated for this
case. It has the second sentence added:

> "If you have received an invitation email, please follow the invitation link
> to get access to a project. If this problem persists please contact support
> to get your account configured."

A user might have actually followed the invitation link and still have no
project assigned.

Fixes #68.