Fixed broken link for MariaDB Java Connector
init.sh script crashed when trying to retrieve the connector from the
AARNET mirror. The AARNET mirror no longer provides the Java Connector
package.

Changed the mirror to the MariaDB official archive and changed the
the SHA256sum to MD5 checksum as the archive only provides the MD5
checksum for said files.

* init.sh no longer crashes when trying to download the connector